lab-ai-gold-completo/lab-ai-gold/apps/api/requirements.txt
# LAB AI GOLD - Backend Dependencies
# Python 3.11

# Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Banco de dados
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
geoalchemy2==0.15.1

# Validação e settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Geoespacial
shapely==2.0.4
pyproj==3.6.1
fastkml==1.0.0
lxml==5.2.1

# Storage
boto3==1.34.93

# Utilitários
python-dotenv==1.0.1
httpx==0.27.0
orjson==3.10.3

# PDF
reportlab==4.1.0
